+++ src/components/fields/AutosuggestField.js
@@ -109,12 +109,13 @@ export default class AutosuggestField extends React.Component {
 		let options = {
 			...uiOptions,
 			onSuggestionSelected: this.onSuggestionSelected,
-			onConfirmUnsuggested: this.onConfirmUnsuggested,
-			onInputChange: this.onInputChange,
+			onUnsuggestedSelected: this.onUnsuggestedSelected,
+			inputProps: {
+				onChange: this.onInputChange
+			},
 			isValueSuggested: this.isValueSuggested,
 			getSuggestionFromValue: this.getSuggestionFromValue,
 			onInformalTaxonGroupSelected: informalTaxonGroups ? this.onInformalTaxonGroupSelected : undefined,
-			getSuggestionValue: this.getSuggestionValue,
 			informalTaxonGroupsValue: props.formData[informalTaxonGroups],
 			taxonGroupID,
 			placeholder: toggled 
@@ -191,13 +192,6 @@ export default class AutosuggestField extends React.Component {
 		return fieldVal;
 	}
 
-	getSuggestionValue = (suggestion, def) => {
-		const {suggestionValueParse} = this.getActiveOptions(getUiOptions(this.props.uiSchema));
-		return suggestionValueParse
-			? this.getSuggestionReceiverValue(suggestion, suggestionValueParse)
-			: def;
-	}
-
 	onSuggestionSelected = (suggestion, mounted) => {
 		if (suggestion === null) suggestion = undefined;
 
@@ -263,7 +257,7 @@ export default class AutosuggestField extends React.Component {
 		this.props.onChange(formData);
 	}
 
-	onInputChange = (value) => {
+	onInputChange = ({target: {value}}) => {
 		let {formData, uiSchema} = this.props;
 		const {inputTransformer} = this.getActiveOptions(getUiOptions(uiSchema));
 		if (inputTransformer) {